third_party/WebKit/LayoutTests/fast/html/imports/import-remove-expected.txt

PASS hello.import is non-null.
Node removal should make import deleted.
PASS hello.import is null
Re-inserting it shouldn't help.
PASS hello.import is non-null.
PASS successfullyParsed is true

TEST COMPLETE
